Set and Queue
A Set is a collection type that has no duplicate values.
Note that Set extends the Collection interface and therefore provides the same methods as other Collection data structures do.

There are three concrete implementations for the Set interface:
  1.HashSet
  2.TreeSet
  3.LinkedHashSet

Queues are data structures that allow us to process items in a First-In, First-Out (FIFO) order. The first item placed in a queue is also the first one removed from the queue.